refactor: split the main kubesplit file in several packages
